Alien class created

+++ include/Rect.h
@@ -8,19 +8,18 @@
 class Rect {
 public:
 	Rect();
-	Rect(int, int, int, int);
+	Rect(float,float,float,float);
 
 	~Rect();
 
 	friend std::ostream& operator << (std::ostream& , const Rect&);
-
 
 	Rect AddVec2(const Vec2&);
 	Vec2 GetCenter();
 	double DistanceTo(Rect);
 	bool ContainVec2(const Vec2&);
 
-	int x, y, w, h;
+	float x, y, w, h;
 };
 
 #endif
